+++ src/main/java/joshie/harvest/api/npc/schedule/ScheduleElement.java
@@ -3,9 +3,13 @@ package joshie.harvest.api.npc.schedule;
 import net.minecraft.entity.EntityAgeable;
 
 public abstract class ScheduleElement<E> {
+    private boolean satisfied = false;
+
     public boolean isSatisfied(EntityAgeable npc) {
-        return true;
+        return satisfied;
     }
 
-    public void execute(EntityAgeable npc) {}
+    public void execute(EntityAgeable npc) {
+        satisfied = true;
+    }
 }