+++ src/main/java/joshie/harvest/buildings/HFBuildings.java
@@ -19,6 +19,7 @@ import joshie.harvest.buildings.special.SpecialRuleChurch;
 import joshie.harvest.buildings.special.SpecialRuleFestivals;
 import joshie.harvest.core.base.render.FMLDefinition;
 import joshie.harvest.core.base.render.MeshIdentical;
+import joshie.harvest.core.util.HFTemplate;
 import joshie.harvest.core.util.annotations.HFLoader;
 import joshie.harvest.npcs.HFNPCs;
 import net.minecraft.block.state.IBlockState;
@@ -61,7 +62,6 @@ public class HFBuildings {
     public static final Building SUPERMARKET = registerBuilding("supermarket").setInhabitants(HFNPCs.MILKMAID, HFNPCs.GS_OWNER).setOffset(7, -1, 12).setTickTime(5);
     public static final Building TOWNHALL = registerBuilding("townhall").setSpecialRules(new SpecialRuleBuildings(9)).setInhabitants(HFNPCs.MAYOR, HFNPCs.PRIEST, HFNPCs.DAUGHTER_ADULT, HFNPCs.DAUGHTER_CHILD).setOffset(10, -1, 17);
     public static final Building FESTIVAL_GROUNDS = registerBuilding("festivals", BuildingFestival.class).setSpecialRules(new SpecialRuleFestivals()).setInhabitants(HFNPCs.TRADER).setOffset(14, -1, 32);
-    public static final Building NEWYEARSEVER = registerBuilding("new_years_eve");
 
     public static void preInit() {}
 
@@ -74,7 +74,9 @@ public class HFBuildings {
     public static void init() {
         HFApi.npc.getGifts().addToBlacklist(STRUCTURES, BLUEPRINTS);
         for (Building building: Building.REGISTRY.getValues()) {
-            BuildingRegistry.INSTANCE.getTemplateForBuilding(building).initTemplate();
+            HFTemplate template = BuildingRegistry.INSTANCE.getTemplateForBuilding(building);
+            if (template != null) template.initTemplate();
+            else System.out.println("Failed to load the template for the building: " + building.getRegistryName().toString());
         }
     }